Split HelpMessageLineDefinition and OptionDefinition into impl classes
