+++ embulk-core/src/main/java/org/embulk/deps/cli/CliParser.java
@@ -43,7 +43,7 @@ public abstract class CliParser {
         }
 
         public Builder addHelpMessageLine(final String message) {
-            this.helpLineDefinitions.add(new HelpMessageLineDefinition(message));
+            this.helpLineDefinitions.add(HelpMessageLineDefinition.create(message));
             return this;
         }
